requirements.txt
pyttsx3
SpeechRecognition
openai
google-generativeai
whisper
python-dotenv
